feat(#84): [Tilak|Pooja] separate native code from react native wrapper
* Create Event Emitter Interface
* Separate Wallet code from Wallet Module
