+++ android/src/main/AndroidManifest.xml
@@ -1,5 +1,4 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="io.mosip.tuvali.openid4vpble">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="io.mosip.tuvali.rnModule">
   <uses-permission android:name="android.permission.BLUETOOTH" />
   <uses-permission android:name="android.permission.BLUETOOTH_ADMIN" />
   <uses-permission android:name="android.permission.BLUETOOTH_SCAN" />
