+++ android/src/main/java/io/mosip/tuvali/common/events/Event.kt
@@ -1,0 +1,5 @@
+package io.mosip.tuvali.common.events
+
+interface Event {
+  val type: String
+}
