+++ android/src/main/java/io/mosip/tuvali/common/events/EventWithoutArgs.kt
@@ -1,0 +1,4 @@
+package io.mosip.tuvali.common.events
+
+interface EventWithoutArgs: Event {
+}
