+++ android/src/main/java/io/mosip/tuvali/common/events/IEventEmitter.kt
@@ -1,0 +1,9 @@
+package io.mosip.tuvali.common.events
+
+import io.mosip.tuvali.exception.ErrorCode
+
+interface IEventEmitter {
+  fun emitEventWithoutArgs(event: EventWithoutArgs)
+  fun emitEventWithArgs(event: EventWithArgs)
+  fun emitError(message: String, code: ErrorCode)
+}
