+++ android/src/main/java/io/mosip/tuvali/common/events/withoutArgs/ConnectedEvent.kt
@@ -1,0 +1,7 @@
+package io.mosip.tuvali.common.events.withoutArgs
+
+import io.mosip.tuvali.common.events.EventWithoutArgs
+
+internal class ConnectedEvent: EventWithoutArgs {
+  override val type = "onConnected"
+}
