+++ android/src/main/java/io/mosip/tuvali/common/events/withoutArgs/DataSentEvent.kt
@@ -1,0 +1,7 @@
+package io.mosip.tuvali.common.events.withoutArgs
+
+import io.mosip.tuvali.common.events.EventWithoutArgs
+
+class DataSentEvent: EventWithoutArgs {
+  override val type = "onDataSent"
+}
