+++ android/src/main/java/io/mosip/tuvali/common/events/withoutArgs/DisconnectedEvent.kt
@@ -1,0 +1,7 @@
+package io.mosip.tuvali.common.events.withoutArgs
+
+import io.mosip.tuvali.common.events.EventWithoutArgs
+
+class DisconnectedEvent: EventWithoutArgs {
+  override val type = "onDisconnected"
+}
