+++ android/src/main/java/io/mosip/tuvali/common/events/withoutArgs/SecureChannelEstablishedEvent.kt
@@ -1,0 +1,7 @@
+package io.mosip.tuvali.common.events.withoutArgs
+
+import io.mosip.tuvali.common.events.EventWithoutArgs
+
+class SecureChannelEstablishedEvent: EventWithoutArgs {
+  override val type = "onSecureChannelEstablished"
+}
