+++ android/src/main/java/io/mosip/tuvali/common/version/VersionDetails.kt
@@ -1,0 +1,7 @@
+package io.mosip.tuvali.common.version
+
+class VersionDetails {
+  companion object {
+    var tuvaliVersion: String = "unknown"
+  }
+}
