+++ android/src/main/java/io/mosip/tuvali/exception/handlers/ExceptionHandler.kt
@@ -1,7 +1,6 @@
 package io.mosip.tuvali.exception.handlers
 
 import android.util.Log
-import com.facebook.react.bridge.Callback
 import io.mosip.tuvali.exception.BLEException
 import io.mosip.tuvali.exception.ErrorCode
 import io.mosip.tuvali.exception.ExceptionUtils
@@ -11,7 +10,7 @@ import io.mosip.tuvali.transfer.Util
 import io.mosip.tuvali.verifier.exception.VerifierException
 import io.mosip.tuvali.wallet.exception.WalletException
 
-class OpenIdBLEExceptionHandler(private val sendError: (String, ErrorCode) -> Unit, private val stopBle: (() -> Unit) -> Unit) {
+class ExceptionHandler(private val sendError: (String, ErrorCode) -> Unit, private val stopBle: (() -> Unit) -> Unit) {
   private val logTag = Util.getLogTag(javaClass.simpleName)
   private var walletExceptionHandler: WalletExceptionHandler = WalletExceptionHandler(sendError)
   private var verifierExceptionHandler: VerifierExceptionHandler = VerifierExceptionHandler(sendError)
