--- android/src/main/java/io/mosip/tuvali/openid4vpble/events/Event.kt
@@ -1,5 +1,0 @@
-package io.mosip.tuvali.openid4vpble.events
-
-interface Event {
-  val type: String
-}
