--- android/src/main/java/io/mosip/tuvali/openid4vpble/events/EventWithoutArgs.kt
@@ -1,4 +1,0 @@
-package io.mosip.tuvali.openid4vpble.events
-
-interface EventWithoutArgs: Event {
-}
