--- android/src/main/java/io/mosip/tuvali/openid4vpble/events/withoutArgs/ConnectedEvent.kt
@@ -1,7 +1,0 @@
-package io.mosip.tuvali.openid4vpble.events.withoutArgs
-
-import io.mosip.tuvali.openid4vpble.events.EventWithoutArgs
-
-internal class ConnectedEvent: EventWithoutArgs {
-  override val type = "onConnected"
-}
