--- android/src/main/java/io/mosip/tuvali/openid4vpble/events/withoutArgs/DataSentEvent.kt
@@ -1,7 +1,0 @@
-package io.mosip.tuvali.openid4vpble.events.withoutArgs
-
-import io.mosip.tuvali.openid4vpble.events.EventWithoutArgs
-
-class DataSentEvent: EventWithoutArgs {
-  override val type = "onDataSent"
-}
