--- android/src/main/java/io/mosip/tuvali/openid4vpble/events/withoutArgs/DisconnectedEvent.kt
@@ -1,7 +1,0 @@
-package io.mosip.tuvali.openid4vpble.events.withoutArgs
-
-import io.mosip.tuvali.openid4vpble.events.EventWithoutArgs
-
-class DisconnectedEvent: EventWithoutArgs {
-  override val type = "onDisconnected"
-}
