--- android/src/main/java/io/mosip/tuvali/openid4vpble/events/withoutArgs/SecureChannelEstablishedEvent.kt
@@ -1,7 +1,0 @@
-package io.mosip.tuvali.openid4vpble.events.withoutArgs
-
-import io.mosip.tuvali.openid4vpble.events.EventWithoutArgs
-
-class SecureChannelEstablishedEvent: EventWithoutArgs {
-  override val type = "onSecureChannelEstablished"
-}
