+++ android/src/main/java/io/mosip/tuvali/rnModule/RNWalletModule.kt
@@ -1,0 +1,35 @@
+package io.mosip.tuvali.rnModule
+
+import com.facebook.react.bridge.ReactApplicationContext
+import com.facebook.react.bridge.ReactContextBaseJavaModule
+import com.facebook.react.bridge.ReactMethod
+import io.mosip.tuvali.wallet.Wallet
+
+class RNWalletModule(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(reactContext) {
+  private val eventEmitter = EventEmitter(reactContext)
+  private var wallet: Wallet = Wallet(eventEmitter, reactContext)
+
+  @ReactMethod(isBlockingSynchronousMethod = true)
+  fun startConnection(uri: String) {
+    wallet.startConnection(uri)
+  }
+
+  @ReactMethod
+  fun sendData(payload: String) {
+    wallet.sendData(payload)
+  }
+
+  @ReactMethod(isBlockingSynchronousMethod = true)
+  fun disconnect() {
+   wallet.disconnect()
+  }
+
+  override fun getName(): String {
+    return NAME
+  }
+
+  companion object {
+    var tuvaliVersion: String = "unknown"
+    const val NAME = "WalletModule"
+  }
+}
