+++ android/src/main/java/io/mosip/tuvali/verifier/IVerifier.kt
@@ -1,0 +1,7 @@
+package io.mosip.tuvali.verifier
+
+interface IVerifier {
+  fun startAdvertisement(advIdentifier: String): String
+  fun disconnect()
+  fun sendVerificationStatus(status: String)
+}
