+++ android/src/main/java/io/mosip/tuvali/wallet/IWallet.kt
@@ -1,0 +1,7 @@
+package io.mosip.tuvali.wallet
+
+interface IWallet {
+  fun startConnection(uri: String)
+  fun sendData(payload: String)
+  fun disconnect()
+}
