+++ src/index.tsx
@@ -1,5 +1,5 @@
 import { NativeEventEmitter, NativeModules, Platform } from 'react-native';
-import type { Verifier, Wallet } from './types/bleshare';
+import type { Verifier, VersionModule, Wallet } from './types/bleshare';
 import { tuvaliVersion } from './tuvaliVersion';
 import { EventTypes, VerificationStatus } from './types/events';
 
@@ -11,6 +11,17 @@ const LINKING_ERROR =
 
 const VERIFIER_NOT_IMPLEMENTATED_ERROR = `Verifier is not yet implemented on IOS. Please remove Verifier usage on IOS Platform`;
 const isIOS = Platform.OS === 'ios';
+
+const versionModule: VersionModule = NativeModules.VersionModule
+  ? NativeModules.VersionModule
+  : new Proxy(
+      {},
+      {
+        get() {
+          throw new Error(LINKING_ERROR);
+        },
+      }
+    );
 
 const wallet: Wallet = NativeModules.WalletModule
   ? NativeModules.WalletModule
@@ -37,6 +48,8 @@ let verifier: Verifier = NativeModules.VerifierModule
       }
     );
 
+versionModule.setTuvaliVersion(tuvaliVersion);
+
 if (!isIOS) {
   setupModule(verifier);
 }
@@ -60,8 +73,6 @@ setupModule(wallet);
 // });
 
 function setupModule(module: any) {
-  module.setTuvaliVersion(tuvaliVersion);
-
   if (Platform.OS === 'android') {
     const eventEmitter = new NativeEventEmitter();
     module.handleDataEvents = (callback: (event: any) => void) =>
